refactor: Move MediaPermissionService to a nested "accessControl" namespace under "model"
